+++ bageldb-flutter/lib/src/bage_db_shared_prefs.dart
@@ -1,0 +1,35 @@
+import 'package:shared_preferences/shared_preferences.dart';
+
+class SP {
+  late SharedPreferences instance;
+
+  Future<SharedPreferences> init() async {
+    instance = await SharedPreferences.getInstance();
+    return instance;
+  }
+
+  Map<String, dynamic> get(String key) {
+    final res = <String, dynamic>{};
+    String? str = instance.getString(key);
+    if (str == null) return res;
+    for (String keyValue in str.split(';;;')) {
+      List<String> kv = keyValue.split(':::');
+      String k = kv[0];
+      String v = kv[1];
+      res[k] = v;
+    }
+    return res;
+  }
+
+  delete(String key) async {
+    await instance.remove(key);
+  }
+
+  set(String key, Map<String, dynamic> value) {
+    List<String> vals = [];
+    for (var k in value.keys) {
+      vals.add('$k:::${value[k]}');
+    }
+    instance.setString(key, vals.join(';;;'));
+  }
+}